refactor: rename list management components
